requirements.txt
python-digitalocean
ansible_runner
emoji
colorama
argparse